# src/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Manage Blockchain",
        "title": "Disashop Blockchain Management API",
        "contact": {
            "name": "Disashop",
            "url": "https://disashop.com",
            "email": "eduardo.alonso@disashop.com"
        },
        "version": "1.0"
    },
    "host": "localhost:30000",
    "basePath": "/api/v1",
    "paths": {
        "/getbalance": {
            "post": {
                "description": "Obtain balance from a given address that belong to address",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Obtain balance of a wallet",
                "parameters": [
                    {
                        "description": "balance",
                        "name": "getbalance",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Balance"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "check health about the service to monitor status",
                "summary": "Check health about the service to monitor status",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/sendtransaction": {
            "post": {
                "description": "send amount to user from wallet to wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Send amount to other user",
                "parameters": [
                    {
                        "description": "transaction",
                        "name": "sendtransaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.SendTransaction"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Balance": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                }
            }
        },
        "api.SendTransaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "from": {
                    "type": "string"
                },
                "mineNow": {
                    "type": "boolean",
                    "default": false
                },
                "to": {
                    "type": "string"
                }
            }
        }
    }
}